[Refactor] group의 Location 삭제 및 관련 로직 수정
