--- src/main/java/com/example/bookiibookii/domain/group/entity/Location.java
@@ -1,30 +1,0 @@
-package com.example.bookiibookii.domain.group.entity;
-
-import com.example.bookiibookii.global.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "location")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Builder
-public class Location extends BaseEntity {
-
-    @Id
-    @Column(name = "location_id")
-    private String id;
-
-    @Column(name = "place_name")
-    private String placeName;
-
-    @Column(name = "address")
-    private String address;
-
-    @Column(name = "address_detail")
-    private String addressDetail;
-
-    @Column(name = "zip_code", length = 10)
-    private String zipCode;
-}
