--- src/main/java/com/example/bookiibookii/domain/group/repository/LocationRepository.java
@@ -1,9 +1,0 @@
-package com.example.bookiibookii.domain.group.repository;
-
-import com.example.bookiibookii.domain.group.entity.Location;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface LocationRepository extends JpaRepository<Location, String> {
-}
